+++ MercadoPagoSDK/MercadoPagoSDK/Core/Checkout/ViewModel/MercadoPagoCheckoutViewModel.swift
@@ -84,7 +84,7 @@ class MercadoPagoCheckoutViewModel: NSObject, NSCopying {
 
     // Payment plugin
     var paymentPlugin: PXSplitPaymentProcessor?
-    var paymentFlow: PXPaymentFlow?
+    private(set) var paymentFlow: PXPaymentFlow?
 
     // Discount and charges
     var chargeRules: [PXPaymentTypeChargeRule]?
@@ -95,6 +95,8 @@ class MercadoPagoCheckoutViewModel: NSObject, NSCopying {
 
     // OneTap Flow
     var onetapFlow: OneTapFlow?
+
+    var postPaymentNotificationName: Notification.Name?
 
     lazy var pxNavigationHandler: PXNavigationHandler = PXNavigationHandler.getDefault()
 
@@ -128,6 +130,10 @@ class MercadoPagoCheckoutViewModel: NSObject, NSCopying {
 
     func setNavigationHandler(handler: PXNavigationHandler) {
         pxNavigationHandler = handler
+    }
+
+    func setPostPaymentNotification(postPaymentNotificationName: Notification.Name) {
+        self.postPaymentNotificationName = postPaymentNotificationName
     }
 
     func hasError() -> Bool {
@@ -738,15 +744,28 @@ extension MercadoPagoCheckoutViewModel {
 extension MercadoPagoCheckoutViewModel {
     func createPaymentFlow(paymentErrorHandler: PXPaymentErrorHandlerProtocol) -> PXPaymentFlow {
         guard let paymentFlow = paymentFlow else {
-            let paymentFlow = PXPaymentFlow(paymentPlugin: paymentPlugin, mercadoPagoServices: mercadoPagoServices, paymentErrorHandler: paymentErrorHandler, navigationHandler: pxNavigationHandler, amountHelper: amountHelper, checkoutPreference: checkoutPreference, ESCBlacklistedStatus: search?.configurations?.ESCBlacklistedStatus)
-            if let productId = advancedConfig.productId {
-                paymentFlow.setProductIdForPayment(productId)
-            }
-            self.paymentFlow = paymentFlow
-            return paymentFlow
+            return buildPaymentFlow(with: paymentErrorHandler)
         }
         paymentFlow.model.amountHelper = amountHelper
         paymentFlow.model.checkoutPreference = checkoutPreference
+        return paymentFlow
+    }
+
+    func buildPaymentFlow(with paymentErrorHandler: PXPaymentErrorHandlerProtocol) -> PXPaymentFlow {
+        let paymentFlow = PXPaymentFlow(
+            paymentPlugin: paymentPlugin,
+            mercadoPagoServices: mercadoPagoServices,
+            paymentErrorHandler: paymentErrorHandler,
+            navigationHandler: pxNavigationHandler,
+            amountHelper: amountHelper,
+            checkoutPreference: checkoutPreference,
+            ESCBlacklistedStatus: search?.configurations?.ESCBlacklistedStatus
+        )
+        if let productId = advancedConfig.productId {
+            paymentFlow.setProductIdForPayment(productId)
+        }
+        paymentFlow.model.postPaymentNotificationName = postPaymentNotificationName
+        self.paymentFlow = paymentFlow
         return paymentFlow
     }
 }