+++ MercadoPagoSDK/MercadoPagoSDK/Services/LocationService.swift
@@ -19,4 +19,3 @@ class LocationService {
     return locationEnabled
   }
 }
-