+++ MercadoPagoSDK/MercadoPagoSDK/UI/AnimatedButton/PXAnimatedButton.swift
@@ -2,6 +2,17 @@ import Foundation
 import MLUI
 
 class PXAnimatedButton: UIButton {
+    private enum Constants {
+        static let defaultAnimationDuration = 0.5
+        static let defaultTimeOut = 15.0
+        static let explosionAnimationDuration = 0.3
+        static let showIconAnimationDuration = 0.6
+        static let hideIconAnimationDuration = 0.4
+        static let scaleFactorForIconAnimation = CGFloat(0.40)
+        static let xScaleForIconAnimation = CGFloat(1.0)
+        static let xScaleForExpandAnimation = CGFloat(50)
+    }
+
     weak var animationDelegate: PXAnimatedButtonDelegate?
     var progressView: ProgressView?
     var status: Status = .normal
@@ -43,19 +54,25 @@ class PXAnimatedButton: UIButton {
 
 // MARK: Animations
 extension PXAnimatedButton: ProgressViewDelegate, CAAnimationDelegate {
-    func startLoading(timeOut: TimeInterval = 15.0) {
+    func startLoading(timeOut: TimeInterval = Constants.defaultTimeOut) {
         progressView = ProgressView(forView: self, loadingColor: #colorLiteral(red: 0.2666666667, green: 0.2666876018, blue: 0.2666300237, alpha: 0.4), timeOut: timeOut)
         progressView?.progressDelegate = self
         setTitle(loadingText, for: .normal)
         status = .loading
     }
 
-    func finishAnimatingButton(color: UIColor, image: UIImage?) {
+    func finishAnimatingButton(color: UIColor, image: UIImage?, interrupt: Bool) {
         status = .expanding
 
         progressView?.doComplete(completion: { [weak self] _ in
             guard let self = self,
                 let anchorView = self.anchorView() else { return }
+
+            if interrupt {
+                self.setTitle("", for: .normal)
+                PXNotificationManager.Post.didFinishButtonAnimation()
+                return
+            }
 
             let animatedViewOriginInAnchorViewCoordinates = self.convert(CGPoint.zero, to: anchorView)
             let animatedViewFrameInAnchorViewCoordinates = CGRect(origin: animatedViewOriginInAnchorViewCoordinates, size: self.frame.size)
@@ -70,19 +87,23 @@ extension PXAnimatedButton: ProgressViewDelegate, CAAnimationDelegate {
             self.animatedView = animatedView
             self.alpha = 0
 
-            let toCircleFrame = CGRect(x: animatedViewFrameInAnchorViewCoordinates.midX - animatedViewFrameInAnchorViewCoordinates.height / 2,
-                                       y: animatedViewFrameInAnchorViewCoordinates.minY,
-                                       width: animatedViewFrameInAnchorViewCoordinates.height,
-                                       height: animatedViewFrameInAnchorViewCoordinates.height)
-
-            let transitionAnimator = UIViewPropertyAnimator(duration: 0.5, dampingRatio: 1, animations: {
+            let toCircleFrame = CGRect(
+                x: animatedViewFrameInAnchorViewCoordinates.midX - animatedViewFrameInAnchorViewCoordinates.height / 2,
+                y: animatedViewFrameInAnchorViewCoordinates.minY,
+                width: animatedViewFrameInAnchorViewCoordinates.height,
+                height: animatedViewFrameInAnchorViewCoordinates.height
+            )
+            let transitionAnimator = UIViewPropertyAnimator(
+                duration: Constants.defaultAnimationDuration,
+                dampingRatio: 1
+            ) {
                 animatedView.frame = toCircleFrame
                 animatedView.layer.cornerRadius = toCircleFrame.height / 2
-            })
+            }
 
-            transitionAnimator.addCompletion({ [weak self] _ in
+            transitionAnimator.addCompletion { [weak self] _ in
                 self?.explosion(color: color, newFrame: toCircleFrame, image: image)
-            })
+            }
 
             transitionAnimator.startAnimation()
         })
@@ -91,40 +112,74 @@ extension PXAnimatedButton: ProgressViewDelegate, CAAnimationDelegate {
     private func explosion(color: UIColor, newFrame: CGRect, image: UIImage?) {
         guard let animatedView = self.animatedView else { return }
 
-        UIView.animate(withDuration: 0.3, animations: { [weak self] in
-            self?.progressView?.alpha = 0
-            animatedView.backgroundColor = color
-        }, completion: { _ in
-            let scaleFactor: CGFloat = 0.40
-            let iconImage = UIImageView(frame: CGRect(x: newFrame.width / 2 - (newFrame.width * scaleFactor) / 2, y: newFrame.width / 2 - (newFrame.width * scaleFactor) / 2, width: newFrame.width * scaleFactor, height: newFrame.height * scaleFactor))
-
-            iconImage.image = image
-            iconImage.contentMode = .scaleAspectFit
-            iconImage.alpha = 0
-
-            animatedView.addSubview(iconImage)
-
-            PXFeedbackGenerator.successNotificationFeedback()
-
-            UIView.animate(withDuration: 0.6, animations: {
-                iconImage.alpha = 1
-                iconImage.transform = CGAffineTransform(scaleX: 1.0, y: 1.0)
+        UIView.animate(
+            withDuration: Constants.explosionAnimationDuration,
+            animations: { [weak self] in
+                self?.progressView?.alpha = 0
+                animatedView.backgroundColor = color
             }, completion: { _ in
-                UIView.animate(withDuration: 0.4, animations: {
-                    iconImage.alpha = 0
-                }, completion: { [weak self] _ in
-                    guard let self = self else { return }
-                    self.superview?.layer.masksToBounds = false
+                PXFeedbackGenerator.successNotificationFeedback()
+                self.iconAnimation(newFrame: newFrame, image: image) {
                     self.animationDelegate?.expandAnimationInProgress()
-                    UIView.animate(withDuration: 0.5, animations: {
-                        animatedView.transform = CGAffineTransform(scaleX: 50, y: 50)
-                    }, completion: { [weak self] _ in
-                        self?.animationDelegate?.didFinishAnimation()
-                    })
-                })
-            })
-        })
+                    self.expandAnimation {
+                        self.animationDelegate?.didFinishAnimation()
+                    }
+                }
+            }
+        )
     }
+
+    private func iconAnimation(newFrame: CGRect, image: UIImage?, completion: @escaping () -> Void) {
+        guard let animatedView = self.animatedView else { return }
+        let scaleFactor: CGFloat = Constants.scaleFactorForIconAnimation
+        let iconImage = UIImageView(
+            frame: CGRect(
+                x: newFrame.width / 2 - (newFrame.width * scaleFactor) / 2,
+                y: newFrame.width / 2 - (newFrame.width * scaleFactor) / 2,
+                width: newFrame.width * scaleFactor,
+                height: newFrame.height * scaleFactor
+            )
+        )
+
+        iconImage.image = image
+        iconImage.contentMode = .scaleAspectFit
+        iconImage.alpha = 0
+        animatedView.addSubview(iconImage)
+
+        UIView.animate(
+            withDuration: Constants.showIconAnimationDuration,
+            animations: {
+                iconImage.alpha = 1
+                iconImage.transform = CGAffineTransform(
+                    scaleX: Constants.xScaleForIconAnimation,
+                    y: Constants.xScaleForIconAnimation
+                )
+            }, completion: { _ in
+                UIView.animate(
+                    withDuration: Constants.hideIconAnimationDuration,
+                    animations: {
+                        iconImage.alpha = 0
+                    }, completion: { _ in completion() }
+                )
+            }
+        )
+    }
+
+        private func expandAnimation(completion: @escaping () -> Void) {
+            guard let animatedView = self.animatedView else { return }
+
+            self.superview?.layer.masksToBounds = false
+            UIView.animate(
+                withDuration: Constants.defaultAnimationDuration,
+                animations: {
+                    animatedView.transform = CGAffineTransform(
+                        scaleX: Constants.xScaleForExpandAnimation,
+                        y: Constants.xScaleForExpandAnimation
+                    )
+                },
+                completion: { _ in completion() }
+            )
+        }
 
     func didFinishProgress() {
         progressView?.doReset()
@@ -217,7 +272,7 @@ extension PXAnimatedButton {
         if let notificationObject = sender.object as? PXAnimatedButtonNotificationObject {
             let image = ResourceManager.shared.getBadgeImageWith(status: notificationObject.status, statusDetail: notificationObject.statusDetail, clearBackground: true)
             let color = ResourceManager.shared.getResultColorWith(status: notificationObject.status, statusDetail: notificationObject.statusDetail)
-            finishAnimatingButton(color: color, image: image)
+            finishAnimatingButton(color: color, image: image, interrupt: notificationObject.interrupt)
         }
     }
 }